Add the core code for the application
I appreciate it's a pretty monumental commit, but I really wasn't sure
of the approach when I first started writing the code, so I built it and
built it, and before I knew it, it was too far along to break it down
into a small commit. So, for one time in this project, there's a
monumental commit.