Move DbTest class database setup to a dedicated trait
This change moves the common code for setting up and tearing down the
test database (using Phinx) to a common, reusable trait. It also
refactors it slightly. I'm, honestly, not sure if this is the best
approach, but it makes the most sense to me as I sit here right now.