Add the templates path in the ConfigProvider
This change register the package's templates in the ConfigProvider,
which in turn will make them broadly available to the application that
uses the package. The prefix had to change to avoid a namespace
collision with any app that uses the package. I don't like the prefix,
but it's a start.