Autoload the database seeders properly
Because of how Phinx loads its database seeders, I couldn't autoload
them using a PSR-0/4 namespace as I normally would. Given that, and me
being just generally unfamiliar with other autoloading options, I didn't
know how to autoload them, until I searched through the Composer
documentation and re-read the classmap autoloading docu'mentation.
So, with the change in this PR, the classes can be used in the tests and
not do anything fancy with using them with Phinx as it normally expects
them to be set up.